requirements.txt
alembic==1.4.2
appdirs==1.4.4
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
distlib==0.3.1
dominate==2.5.1
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-User==1.0.2.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.3
passlib==1.7.2
Pillow==7.1.1
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.15
virtualenv==20.0.18
visitor==0.1.3
Werkzeug==1.0.1
Whoosh==2.7.4
WTForms==2.2.1
